// frontend/app/generated/models/CategoryResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Borsibaar API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CategoryResponse
 */
export interface CategoryResponse {
    /**
     * 
     * @type {number}
     * @memberof CategoryResponse
     */
    id: number;
    /**
     * 
     * @type {string}
     * @memberof CategoryResponse
     */
    name: string;
    /**
     * 
     * @type {boolean}
     * @memberof CategoryResponse
     */
    dynamicPricing?: boolean;
}

/**
 * Check if a given object implements the CategoryResponse interface.
 */
export function instanceOfCategoryResponse(value: object): value is CategoryResponse {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    return true;
}

export function CategoryResponseFromJSON(json: any): CategoryResponse {
    return CategoryResponseFromJSONTyped(json, false);
}

export function CategoryResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): CategoryResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'name': json['name'],
        'dynamicPricing': json['dynamicPricing'] == null ? undefined : json['dynamicPricing'],
    };
}

export function CategoryResponseToJSON(json: any): CategoryResponse {
    return CategoryResponseToJSONTyped(json, false);
}

export function CategoryResponseToJSONTyped(value?: CategoryResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'dynamicPricing': value['dynamicPricing'],
    };
}