// frontend/app/generated/models/InventoryResponse.ts
export interface InventoryResponse {
    /**
     * 
     * @type {number}
     * @memberof InventoryResponse
     */
    id: number;
    /**
     * 
     * @type {number}
     * @memberof InventoryResponse
     */
    organizationId: number;
    /**
     * 
     * @type {number}
     * @memberof InventoryResponse
     */
    productId: number;
    /**
     * 
     * @type {string}
     * @memberof InventoryResponse
     */
    productName: string;
    /**
     * 
     * @type {number}
     * @memberof InventoryResponse
     */
    quantity: number;
    /**
     * 
     * @type {number}
     * @memberof InventoryResponse
     */
    unitPrice: number;
    /**
     * 
     * @type {string}
     * @memberof InventoryResponse
     */
    description: string;
    /**
     * 
     * @type {number}
     * @memberof InventoryResponse
     */
    basePrice: number;
    /**
     * 
     * @type {number}
     * @memberof InventoryResponse
     */
    minPrice: number;
    /**
     * 
     * @type {number}
     * @memberof InventoryResponse
     */
    maxPrice: number;
    /**
     * 
     * @type {Date}
     * @memberof InventoryResponse
     */
    updatedAt: Date;
}